+++ src/contexts/navigation/types.js
@@ -1,12 +1,14 @@
 export const COLLAPSE = 'collapse' ;
 export const DIVIDER  = 'divider'  ;
 export const LINK     = 'link'     ;
+export const TITLE    = 'title'    ;
 
 const types =
 [
     COLLAPSE ,
     DIVIDER ,
-    LINK
+    LINK ,
+    TITLE
 ]
 
 export default types ;