+++ src/display/ui/navigation/Items.jsx
@@ -6,11 +6,12 @@
  * @module display/ui/navigation/items
  */
 
-import { COLLAPSE , DIVIDER , LINK } from '../../../contexts/navigation/types' ;
+import { COLLAPSE , DIVIDER , LINK , TITLE } from '../../../contexts/navigation/types' ;
 
 import Collapse from './Collapse' ;
 import Divider  from './Divider' ;
 import Link     from './Link' ;
+import Title    from './Title' ;
 
 /**
  * @type {Record<string, React.ComponentType>}
@@ -20,6 +21,7 @@ const items =
     [ COLLAPSE ] : Collapse ,
     [ DIVIDER  ] : Divider ,
     [ LINK     ] : Link ,
+    [ TITLE    ] : Title ,
 } ;
 
 export default items ;